Update to Connection .
Added a token validator, to validate our MOD status across channels. If we are modded in all channels, Twitch softens the ratelimits. If not the standard limits apply.

Fix to User.

Still working on command system and finishing up the ACTIONS.

TODO: Add a RateLimit validator, and Integration Startup.

Integration startup will allow for ease of use in other Async Libs like, an aiohttp web server or discord.py.