Added `http._get_streams`.
This method is dynamic in that it allows you to provide the Twitch maximum amount of 100 either names or ids, and returns the data as a preassembled dictionary.

Adjusted the handling for Headers. At this point this basically does nothing but provide a base for future API endpoint implementation.

Added client version of `http._get_streams`.

Fixed some documentation.

Adjusted client error for future Execeptions.

Signed-off-by: MysterialPy <mysterialpy@gmail.com>