Update websocket to allow for a properly timed ready event.
The WebSocket has been overhauled, to allow for the `event_ready` to be
fired at the correct time. Many changes have been made including,
changes to `join_channels`, `process_data`, `auth_seq`, `_listen`.

`event_ready` should now only fire once the channel cache has been
verified complete and you have authenticated successfully on the
twitch servers.

A `TimeoutError` will be raised on any channels that the bot is unable
to JOIN. Since Twitch doesn't send any notification whatsoever of invalid
channels, this is done by adding a timeout to a wait and `Future`. If
the `Future` does not result from the JOIN action, it will raise after
10 seconds have passed.

Logging has been added in many places. This needs to be monitored closely.
PRIVMSG was chosen to be excluded from logging as it is a very high frequency
action.