Added support for `USERNOTICE` events.
Two new events have been added to the bot. Alongside a new dataclass
`NoticeSubscription`.

`event_usernotice` and a sub-event `event_usernotice_sub` add functionality
for receiving USERNOTICE events.

The former takes in a sole parameter `tags`, which is a dict containing
all the relevant info to the USERNOTICE event received by Twitch.

The latter takes in a sole parameter `metadata`, which is the new dataclass
`NoticeSubscription` and contains all the relevant info to the subscription
event, including, a Channel object, a User object, the tags dict, and extra
information relevant to the subscription.

For more information on USERNOTICE's please see:
https://dev.twitch.tv/docs/irc/tags/#usernotice-twitch-tags