Add token validation and remove extraneous required arguments.
Previously Client/Bot required you to provide extraneous arguments like nick, client_id and api_token, which was unneeded and confusing.

This commit removes the need for these and allows for a single argument `token` (An Access Token) to be passed, which is then validated, providing you automatically with previously needed arguments.