Add support for PRIVMSG(ECHO).
An ECHO is the message the bot sent. This will be received as standard message in `event_message` but the messages attribute `.echo` will be set to True. Messages no sent by the bot will have `.echo` set to False.

If you do not wish to receive ECHO messages, you can simply:

`if message.echo: return` in your `event_message`.