Fix message content while handling commands in reply messages (#306)
* Fix message content while handling reply messages

Instead of stripping only the user who was replied to, currently only the second word of the message is set as message content when handling commands. The rest of the message got stripped to. 

So `@dnns01_ This is a reply` was changed to `This` instead of `This is a reply`

* Updated changelog